Adding Markov Chain as script with out directory
The Markov Chain as an end to end script. It outputs images of plots (PNG), models (PKL), predictions and actuals as tuple of two lists (TXT) as files of appropriate extension. It prints out Mean Square Error of the test and training as well as other details for reference. NOTE: When comparing MSE to other models, node that this model in its current form is only making predictions one month ahead of what it can see as current and previous "states" (load values of prev and cur).